+++ lib/pages/savepage.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:handrange/creategraph.dart';
-import 'package:handrange/drawer.dart';
-import 'package:handrange/sql.dart';
+import '../functions/creategraph.dart';
+import '../compornents/drawer.dart';
+import '../sqls/sql.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
-import 'package:handrange/light.dart';
+import '../providers/light.dart';
 
 class MyApp extends StatelessWidget {
   @override
